\name{quickTrends-package}
\alias{quickTrends-package}
\alias{quickTrends}
\docType{package}
\title{
\packageTitle{quickTrends}
}
\description{
\packageDescription{quickTrends}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{quickTrends}
\packageIndices{quickTrends}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{quickTrends}

Maintainer: \packageMaintainer{quickTrends}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
